Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 1590-make-sum-divisible-by-p/1590-make-sum-divisible-by-p.java
@@ -1,13 +1,13 @@
 class Solution {
     public int minSubarray(int[] nums, int p) {
         int n=nums.length;
-        long prefixSum=0;
+        int prefixSum=0;
         // prefixSum
         for(int i=0;i<n;i++){
             prefixSum=prefixSum+nums[i];
 
         }
-        int remainder=(int)(prefixSum%p); //number to delete
+        int remainder=prefixSum%p; //number to delete
         if(remainder==0)return 0;
    
         HashMap<Integer,Integer>map=new HashMap<>();
